+++ share/objects/update-map-information.lua
@@ -32,8 +32,6 @@ if objectexists ('A06Z') then
 			'|c00ff33ff-air|r Displays air rounds.|n' ..
 			'|c00ff33ff-debug|r Attempt to recover from a variety of issues.|n' ..
 			'|c00ff33ff-zoom value|r Sets the camera to the specified value. ' ..
-			'The default Warcraft 3 zoom is `1650`.|n' ..
-			'|c00ff33ff-reminders on|off|r Toggles the status of built-in ' ..
-			'reminders.')
+			'The default Warcraft 3 zoom is `1650`.')
 	end
 end