+++ docs/changelog.md
@@ -5,6 +5,9 @@ full list of changes to the project, please consult the commit log.
 
 ## [1.4.1] - Unreleased
 
+### Removed
+- The `-reminders` command has been removed.
+
 ### Changed
 - Structures are no longer removed from player control during the placement
   and selection phase.